track events to agents, remove one attribute
